--- src/main/java/techcourse/myblog/domain/NoArticleFoundException.java
@@ -1,4 +1,0 @@
-package techcourse.myblog.domain;
-
-public class NoArticleFoundException extends RuntimeException {
-}
