--- src/main/resources/templates/head.html
@@ -1,35 +1,0 @@
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, shrink-to-fit=no">
-<title>우아한테크코스 기술블로그</title>
-<meta name="robots" content="index, follow">
-
-<meta name="keyword" content="우아한테크코스, 프로그래밍, 개발자, 백엔드, 프론트엔드, Backend, Frontend, IOS, Android">
-<meta name="description" id="metaDesc" content="우아한테크코스에서 개발자들을 위해 디자인된 교육을 수강해보세요.">
-<meta property="og:type" content="website">
-<meta property="og:title" content="우아한테크코스">
-<meta property="og:description" id="metaOgDesc" content="우아한테크코스 크루 ooo의 개발 블로그">
-<meta property="og:image" content="/images/logo/logo_thumnail_bg.jpg">
-<meta property="og:url" content="https://woowacourse.github.io/">
-
-<meta name="twitter:card" content="website">
-<meta name="twitter:title" content="우아한테크코스">
-<meta name="twitter:description" id="metaTwitterDesc" content="우아한테크코스 크루 ooo의 개발 블로그">
-<meta name="twitter:image" content="/images/logo/logo_thumnail_bg.jpg">
-<meta name="twitter:domain" content="https://woowacourse.github.io/">
-
-<link rel="shortcut icon" href="/images/logo/favicon.ico">
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/lykmapipo/themify-icons@0.1.2/css/themify-icons.css">
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.7.0/animate.min.css">
-<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN" crossorigin="anonymous">
-<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-
-<link rel="stylesheet" href="/css/app.css">
-<link rel="stylesheet" href="/css/customize.css">
-<link rel="stylesheet" href="/css/main.css">
-<link rel="stylesheet" href="/css/index.css">
-<link rel="stylesheet" href="/css/article-edit.css">
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.33.0/codemirror.css">
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.12.0/styles/github.min.css">
-<link rel="stylesheet" href="https://uicdn.toast.com/tui-editor/latest/tui-editor.css">
-<link rel="stylesheet" href="https://uicdn.toast.com/tui-editor/latest/tui-editor-contents.css">
